% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectModels.R
\name{FixedEffect_models}
\alias{FixedEffect_models}
\title{Use FixedEffectModels.jl to run large fixed effect models in julia on multiple models}
\usage{
FixedEffect_models(dt, lhs, rhs, fe = NULL, weights = NULL,
  vcov = NULL)
}
\arguments{
\item{dt}{dataset of interest}

\item{lhs}{String, a vector of Y regression variables}

\item{rhs}{String, a vector of X dependent variables}

\item{fe}{String, a vector of Fixed effects}

\item{weights}{Vector of regression weights (not yet coded)}

\item{vcov}{Vector of specification for the error}
}
\value{
The return value will be a list which contains two elements at this point
  results: includes most of the observation from the julia call
  summary: includes information that is of importance to write a table
}
\description{
\code{FixedEffect_models} returns the results of a linear fixed effect regression
}
\examples{
  # See vignettes and readme
\dontrun{
lhs <- c("log_ewemt", "f1_log_ewemt")
rhs <- c("MP + retail_index + retail_index_MP",
        "MP + discount_rate + discount_rate_MP")
fe  <- c("date_y",
         "date_y:fed_district + quarter:fed_district + industry_code_num:quarter")
vcov <- c("robust", "cluster(date_y)")
}
}
