% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectsetup.R
\name{FixedEffect_setup}
\alias{FixedEffect_setup}
\title{Do initial setup for FixedEffectjlr package.}
\usage{
FixedEffect_setup(...)
}
\arguments{
\item{...}{arguments passed to \code{JuliaCall::julia_setup}.}
}
\description{
\code{FixedEffect_setup} does the initial setup for ipoptjlr package.
}
\examples{
\dontrun{
# FixedEffect_setup()
}

}
