% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectInteract.R
\name{FixedEffectInteract}
\alias{FixedEffectInteract}
\title{Use FixedEffectModelsIV.jl to run large fixed effect models in julia}
\usage{
FixedEffectInteract(dt, lhs, rhs, ife, rank_ife, fe = NULL, weights = NULL,
  vcov = NULL, save_res = FALSE)
}
\arguments{
\item{dt}{dataset of interest}

\item{lhs}{String: lhs variable}

\item{rhs}{String: rhs variables}

\item{ife}{String of interacted fixed effects (id1 + id2)}

\item{rank_ife}{Integer: number of factors to be estimated}

\item{fe}{expression of fixed effects id1 + id2:id3}

\item{weights}{expression of weights}

\item{vcov}{Specification for the error}

\item{save_res}{Save the results of the model}
}
\value{
The return value will be a list which contains two elements at this point
  results: includes most of the observation from the julia call
  summary: includes information that is of importance to write a table
}
\description{
\code{FixedEffectInteract_nse} returns the results of a 2 stage linear fixed effect regression
}
\examples{
# See vignettes and readme
# Short example
\dontrun{
  df = Ecdat::Cigar
  FixedEffectInteract(df, "sales", "price", "state+year", 2, "state", vcov = "robust")
}

}
