% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectIV.R
\name{FixedEffectIV_nse}
\alias{FixedEffectIV_nse}
\title{Use FixedEffectModelsIV.jl to run large fixed effect models in julia}
\usage{
FixedEffectIV_nse(dt, formula, iv, fe = NULL, weights = NULL,
  vcov = NULL, save_res = FALSE, print = TRUE)
}
\arguments{
\item{dt}{dataset of interest}

\item{formula}{formula of the y ~ x1 + x2 type}

\item{iv}{formula endogenous on exogenous variable}

\item{fe}{expression of fixed effects id1 + id2:id3}

\item{weights}{expression of weights}

\item{vcov}{error types, expression either robust or cluster(id1)}

\item{save_res}{Do we save the residuals}

\item{print}{Do we print the results (default is yes)}
}
\value{
The return value will be a list which contains two elements at this point
  results: includes most of the observation from the julia call
  summary: includes information that is of importance to write a table
}
\description{
\code{FixedEffectIV_nse} returns the results of a 2 stage linear fixed effect regression
}
\examples{
  # See vignettes and readme

}
