% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectjlr.R
\docType{package}
\name{FixedEffectjlr-package}
\alias{FixedEffectjlr}
\alias{FixedEffectjlr-package}
\title{FixedEffectjlr a package for Fixed Effect Models}
\description{
FixedEffectjlr uses juliacall and `FixedEffectModels.jl` to perform
  fast high dimensional fixed effect regressions within R
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/eloualiche/FixedEffectjlr}
  \item Report bugs at \url{https://github.com/eloualiche/FixedEffectjlr/issues}
}

}
\author{
\strong{Maintainer}: Erik Loualiche \email{eloualic@umn.edu}

}
