% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectModels.R
\name{FixedEffect}
\alias{FixedEffect}
\title{Use FixedEffectModels.jl to run large fixed effect models in julia}
\usage{
FixedEffect(dt, lhs, rhs, fe = NULL, weights = NULL, vcov = NULL,
  save_res = FALSE, print = TRUE)
}
\arguments{
\item{dt}{dataset of interest}

\item{lhs}{String, Y regression variable}

\item{rhs}{String, X formula}

\item{fe}{Fixed effects}

\item{weights}{Regression weights}

\item{vcov}{Specification for the error}

\item{save_res}{Do we save the residuals}
}
\value{
The return value will be a list which contains two elements at this point
  results: includes most of the observation from the julia call
  summary: includes information that is of importance to write a table
}
\description{
\code{FixedEffect} returns the results of a linear fixed effect regression
}
\examples{
See vignettes and readme

}
